+++ Changelog.md
@@ -37,6 +37,8 @@ Deprecations:
 * Add deprecation warning for `config.filter_run :focused` since
   example aliases `fit` and `focus` will no longer include
   `:focused` metadata but will continue to include `:focus`. (Myron Marston)
+* Deprecate filtering by `:line_number` (e.g. `--line-number` from the
+  CLI). Use location filtering instead. (Myron Marston)
 
 ### 2.99.0.beta2 / 2014-02-17
 [full changelog](http://github.com/rspec/rspec-core/compare/v2.99.0.beta1...v2.99.0.beta2)